Fixed x64 compilation warnings
git-svn-id: http://miranda.googlecode.com/svn/trunk@11370 83df23c2-51fd-11de-bb24-2d5e9b4af41d

+++ miranda/protocols/Gadu-Gadu/image.c
@@ -780,7 +780,7 @@ int gg_img_displayasmsg(GGPROTO *gg, HANDLE hContact, void *img)
 {
 	GGIMAGEENTRY *dat = (GGIMAGEENTRY *)img;
 	char szPath[MAX_PATH], *path = (char *)alloca(MAX_PATH);
-	int tPathLen;
+	size_t tPathLen;
 	FILE *fp;
 
 	if (gg->hImagesFolder == NULL || FoldersGetCustomPath(gg->hImagesFolder, path, MAX_PATH, "")) {
